+++ sdk/src/main/java/com/hedera/hashgraph/sdk/AccountId.java
@@ -58,7 +58,7 @@ public final class AccountId implements Comparable<AccountId> {
     public final PublicKey aliasKey;
 
     @Nullable
-    public final EvmAddress aliasEvmAddress;
+    public final EvmAddress evmAddress;
 
     @Nullable
     private final String checksum;
@@ -98,7 +98,7 @@ public final class AccountId implements Comparable<AccountId> {
         this.num = num;
         this.checksum = checksum;
         this.aliasKey = null;
-        this.aliasEvmAddress = null;
+        this.evmAddress = null;
     }
 
     /**
@@ -115,14 +115,14 @@ public final class AccountId implements Comparable<AccountId> {
         @Nonnegative long num,
         @Nullable String checksum,
         @Nullable PublicKey aliasKey,
-        @Nullable EvmAddress aliasEvmAddress
+        @Nullable EvmAddress evmAddress
     ) {
         this.shard = shard;
         this.realm = realm;
         this.num = num;
         this.checksum = checksum;
         this.aliasKey = aliasKey;
-        this.aliasEvmAddress = aliasEvmAddress;
+        this.evmAddress = evmAddress;
     }
 
     /**
@@ -133,6 +133,9 @@ public final class AccountId implements Comparable<AccountId> {
      * @throws IllegalArgumentException when the account id and checksum are invalid
      */
     public static AccountId fromString(String id) {
+        if ((id.startsWith("0x") && id.length() == 42) || id.length() == 40)
+            return fromEvmAddress(id);
+
         try {
             return EntityIdHelper.fromString(id, AccountId::new);
         } catch (IllegalArgumentException error) {
@@ -157,6 +160,57 @@ public final class AccountId implements Comparable<AccountId> {
     }
 
     /**
+     * Retrieve the account id from an EVM address.
+     *
+     * @param evmAddress                a string representing the EVM address
+     * @return                          the account id object
+     */
+    public static AccountId fromEvmAddress(String evmAddress) {
+        return fromEvmAddress(evmAddress, 0, 0);
+    }
+
+    /**
+     * Retrieve the account id from an EVM address.
+     *
+     * @param evmAddress                a string representing the EVM address
+     * @param shard                     the shard part of the account id
+     * @param realm                     the shard realm of the account id
+     * @return                          the account id object
+     */
+    public static AccountId fromEvmAddress(String evmAddress, @Nonnegative long shard, @Nonnegative long realm) {
+        return fromEvmAddress(EvmAddress.fromString(evmAddress), shard, realm);
+    }
+
+    /**
+     * Retrieve the account id from an EVM address.
+     *
+     * @param evmAddress                an EvmAddress instance
+     * @return                          the account id object
+     */
+    public static AccountId fromEvmAddress(EvmAddress evmAddress) {
+        return fromEvmAddress(evmAddress, 0, 0);
+    }
+
+    /**
+     * Retrieve the account id from an EVM address.
+     *
+     * @param evmAddress                an EvmAddress instance
+     * @param shard                     the shard part of the account id
+     * @param realm                     the shard realm of the account id
+     * @return                          the account id object
+     */
+    public static AccountId fromEvmAddress(EvmAddress evmAddress, @Nonnegative long shard, @Nonnegative long realm) {
+        return new AccountId(
+            shard,
+            realm,
+            0,
+            null,
+            null,
+            evmAddress
+        );
+    }
+
+    /**
      * Retrieve the account id from a solidity address.
      *
      * @param address                   a string representing the address
@@ -173,14 +227,24 @@ public final class AccountId implements Comparable<AccountId> {
      * @return                          the account id object
      */
     static AccountId fromProtobuf(AccountID accountId) {
+        PublicKey aliasKey = null;
+        EvmAddress evmAddress = null;
+
+        if (accountId.hasAlias()) {
+            if (accountId.getAlias().size() == 20) {
+                evmAddress = EvmAddress.fromAliasBytes(accountId.getAlias());
+            } else {
+               aliasKey = PublicKey.fromAliasBytes(accountId.getAlias());
+            }
+        }
         Objects.requireNonNull(accountId);
         return new AccountId(
             accountId.getShardNum(),
             accountId.getRealmNum(),
             accountId.getAccountNum(),
             null,
-            accountId.hasAlias() ? PublicKey.fromAliasBytes(accountId.getAlias()) : null,
-            accountId.hasAlias() ? EvmAddress.fromAliasBytes(accountId.getAlias()) : null
+            aliasKey,
+            evmAddress
         );
     }
 
@@ -215,8 +279,8 @@ public final class AccountId implements Comparable<AccountId> {
             .setRealmNum(realm);
         if (aliasKey != null) {
             accountIdBuilder.setAlias(aliasKey.toProtobufKey().toByteString());
-        } else if (aliasEvmAddress != null) {
-            accountIdBuilder.setAlias(ByteString.copyFrom(aliasEvmAddress.toBytes()));
+        } else if (evmAddress != null) {
+            accountIdBuilder.setAlias(ByteString.copyFrom(evmAddress.toBytes()));
         }else {
             accountIdBuilder.setAccountNum(num);
         }
@@ -240,7 +304,7 @@ public final class AccountId implements Comparable<AccountId> {
      * @throws BadEntityIdException     when the account id and checksum are invalid
      */
     public void validateChecksum(Client client) throws BadEntityIdException {
-        if (aliasKey == null && aliasEvmAddress == null) {
+        if (aliasKey == null && evmAddress == null) {
             EntityIdHelper.validate(shard, realm, num, client, checksum);
         }
     }
@@ -268,8 +332,8 @@ public final class AccountId implements Comparable<AccountId> {
     public String toString() {
         if (aliasKey != null) {
             return "" + shard + "." + realm + "." + aliasKey.toStringDER();
-        } else if (aliasEvmAddress != null) {
-            return "" + shard + "." + realm + "." + aliasEvmAddress.toString();
+        } else if (evmAddress != null) {
+            return "" + shard + "." + realm + "." + evmAddress.toString();
         } else {
             return EntityIdHelper.toString(shard, realm, num);
         }
@@ -282,8 +346,8 @@ public final class AccountId implements Comparable<AccountId> {
      * @return                          the account id with checksum
      */
     public String toStringWithChecksum(Client client) {
-        if (aliasKey != null || aliasEvmAddress != null) {
-            throw new IllegalStateException("toStringWithChecksum cannot be applied to AccountId with aliasKey or aliasEvmAddress");
+        if (aliasKey != null || evmAddress != null) {
+            throw new IllegalStateException("toStringWithChecksum cannot be applied to AccountId with aliasKey or evmAddress");
         } else {
             return EntityIdHelper.toStringWithChecksum(shard, realm, num, client, checksum);
         }
@@ -291,9 +355,16 @@ public final class AccountId implements Comparable<AccountId> {
 
     @Override
     public int hashCode() {
+        byte[] aliasBytes = null;
+
+        if (aliasKey != null) {
+            aliasBytes = aliasKey.toBytes();
+        } else if (evmAddress != null) {
+            aliasBytes = evmAddress.toBytes();
+        }
+
         return Objects.hash(
-            shard, realm, num,
-            (aliasKey != null) ? aliasKey.toBytes() : ((aliasEvmAddress != null) ? aliasEvmAddress.toBytes() : null)
+            shard, realm, num, Arrays.hashCode(aliasBytes)
         );
     }
 
@@ -311,12 +382,12 @@ public final class AccountId implements Comparable<AccountId> {
         if ((aliasKey == null) != (otherId.aliasKey == null)) {
             return false;
         }
-        if ((aliasEvmAddress == null) != (otherId.aliasEvmAddress == null)) {
+        if ((evmAddress == null) != (otherId.evmAddress == null)) {
             return false;
         }
         return shard == otherId.shard && realm == otherId.realm && num == otherId.num &&
             (aliasKey == null || aliasKey.equals(otherId.aliasKey)) &&
-            (aliasEvmAddress == null || aliasEvmAddress.equals(otherId.aliasEvmAddress));
+            (evmAddress == null || evmAddress.equals(otherId.evmAddress));
     }
 
     @Override
@@ -340,12 +411,12 @@ public final class AccountId implements Comparable<AccountId> {
         if (aliasKey != null) {
             return aliasKey.toStringDER().compareTo(o.aliasKey.toStringDER());
         }
-        if ((aliasEvmAddress == null) != (o.aliasEvmAddress == null)) {
-            return aliasEvmAddress != null ? 1 : -1;
+        if ((evmAddress == null) != (o.evmAddress == null)) {
+            return evmAddress != null ? 1 : -1;
         }
-        if (aliasEvmAddress == null) {
+        if (evmAddress == null) {
             return 0;
         }
-        return aliasEvmAddress.toString().compareTo(o.aliasEvmAddress.toString());
+        return evmAddress.toString().compareTo(o.evmAddress.toString());
     }
 }