+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Status.java
@@ -1519,7 +1519,12 @@ public enum Status {
     /**
      * The account cannot be modified. Account's key is not set
      */
-    ACCOUNT_IS_IMMUTABLE(ResponseCodeEnum.ACCOUNT_IS_IMMUTABLE);
+    ACCOUNT_IS_IMMUTABLE(ResponseCodeEnum.ACCOUNT_IS_IMMUTABLE),
+
+    /**
+     * An alias that is assigned to an account or contract cannot be assigned to another account or contract.
+     */
+    ALIAS_ALREADY_ASSIGNED(ResponseCodeEnum.ALIAS_ALREADY_ASSIGNED);
 
     final ResponseCodeEnum code;
 
@@ -1819,6 +1824,7 @@ public enum Status {
             case INSUFFICIENT_BALANCES_FOR_RENEWAL_FEES -> INSUFFICIENT_BALANCES_FOR_RENEWAL_FEES;
             case TRANSACTION_HAS_UNKNOWN_FIELDS -> TRANSACTION_HAS_UNKNOWN_FIELDS;
             case ACCOUNT_IS_IMMUTABLE -> ACCOUNT_IS_IMMUTABLE;
+            case ALIAS_ALREADY_ASSIGNED -> ALIAS_ALREADY_ASSIGNED;
             case UNRECOGNIZED ->
                 // NOTE: Protobuf deserialization will not give us the code on the wire
                 throw new IllegalArgumentException(