+++ sdk/src/test/java/com/hedera/hashgraph/sdk/AccountIdTest.java
@@ -143,7 +143,7 @@ class AccountIdTest {
     }
 
     @Test
-    void fromStringWithAliasEvmAddress() {
+    void fromStringWithEvmAddress() {
         SnapshotMatcher.expect(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toString()).toMatchSnapshot();
     }
 
@@ -163,12 +163,12 @@ class AccountIdTest {
     }
 
     @Test
-    void toBytesAlias() throws InvalidProtocolBufferException {
+    void toBytesAlias() {
         SnapshotMatcher.expect(Hex.toHexString(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82dab5c15ea149f02d34a012087b163516dd70f44acafabf7777").toBytes())).toMatchSnapshot();
     }
 
     @Test
-    void toBytesAliasEvmAddress() {
+    void toBytesEvmAddress() {
         SnapshotMatcher.expect(Hex.toHexString(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toBytes())).toMatchSnapshot();
     }
 
@@ -204,13 +204,20 @@ class AccountIdTest {
     }
 
     @Test
-    void fromBytesAliasEvmAddress() throws InvalidProtocolBufferException {
+    void fromBytesEvmAddress() throws InvalidProtocolBufferException {
         SnapshotMatcher.expect(AccountId.fromBytes(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da").toBytes()).toString()).toMatchSnapshot();
     }
 
     @Test
-    void toFromProtobufAliasEvmAddress() {
+    void toFromProtobufEvmAddress() {
         var id1 = AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82da");
+        var id2 = AccountId.fromProtobuf(id1.toProtobuf());
+        assertThat(id2).isEqualTo(id1);
+    }
+
+    @Test
+    void toFromProtobufRawEvmAddress() {
+        var id1 = AccountId.fromString("302a300506032b6570032100114e6abc371b82da");
         var id2 = AccountId.fromProtobuf(id1.toProtobuf());
         assertThat(id2).isEqualTo(id1);
     }
@@ -218,5 +225,24 @@ class AccountIdTest {
     @Test
     void toSolidityAddress() {
         SnapshotMatcher.expect(new AccountId(5005).toSolidityAddress()).toMatchSnapshot();
+    }
+
+    @Test
+    void fromEvmAddress() {
+        String evmAddress = "302a300506032b6570032100114e6abc371b82da";
+        var id = AccountId.fromEvmAddress(evmAddress, 5, 9);
+
+        assertThat(id.evmAddress).hasToString(evmAddress);
+        assertThat(id.shard).isEqualTo(5);
+        assertThat(id.realm).isEqualTo(9);
+    }
+
+    @Test
+    void fromEvmAddressWithPrefix() {
+        String evmAddressString = "302a300506032b6570032100114e6abc371b82da";
+        EvmAddress evmAddress = EvmAddress.fromString(evmAddressString);
+        var id1 = AccountId.fromEvmAddress(evmAddress);
+        var id2 = AccountId.fromEvmAddress("0x" + evmAddressString);
+        assertThat(id2).isEqualTo(id1);
     }
 }